GameDB_Forum/requirements.txt
Django
djangorestframework==3.15.1
pillow==10.3.0
requests==2.32.3
psycopg2-binary==2.9.6
gunicorn==20.1.0
django-debug-toolbar
